feat: update protocol config

+++ packages/protocol/contracts/shared/data/LibSharedData.sol
@@ -8,6 +8,6 @@ library LibSharedData {
         uint8 sharingPctg;
         uint32 gasIssuancePerSecond;
         uint64 minGasExcess;
-        uint32 maxGasIssuancePerBlock;
+        uint64 maxGasIssuancePerBlock;
     }
 }
